+++ pyportable_installer/prebuild.py
@@ -1,5 +1,5 @@
-import re
 import os
+import re
 import shutil
 from compileall import compile_dir
 from os import path as ospath
@@ -56,13 +56,13 @@ def process_pyproject(pyproj_file):
         return pretty_path(ospath.abspath(f'{pyproj_dir}/{p}'))
     
     def relpath(p: str, start):
-        # 注意: pyproj_dir 与 start 可能是不同的. 在 conf_i 中, 所有相对路径均是指
-        # 相对于 pyproj_dir 的, 而在本函数中, 相对路径的计算是相对于
-        # conf_i['build']['idir'] 的父目录而言 (该目录相当于 `_apply_config:vars
-        # :srcdir`)
-        # 当 pyproject.json 位于要打包的项目源代码的父目录时, pyproj_dir 与 start
-        # 相同 (这是推荐的放置位置); 如果 pyproject.json 放置不当, 比如放在了其他
-        # 目录, 那么就会不同.
+        # 注意: pyproj_dir 与 start 可能是不同的. 在 conf_i 中, 所有相对路径均是
+        # 指相对于 pyproj_dir 的, 而在本函数中, 相对路径的计算是相对于
+        # conf_i['build']['proj_dir'] 的父目录而言 (该目录相当于 `_apply_config
+        # :vars:srcdir`)
+        # 当 pyproject.json 位于要打包的项目源代码的父目录时, pyproj_dir 与
+        # start 相同 (这是推荐的放置位置); 如果 pyproject.json 放置不当, 比如放
+        # 在了其他目录, 那么就会不同.
         if p == '': return ''
         if p[1] == ':': return pretty_path(ospath.relpath(p, start))
         return pretty_path(ospath.relpath(abspath(p), start))
@@ -79,36 +79,40 @@ def process_pyproject(pyproj_file):
     if pyver := conf_i['build']['required']['python_version']:
         assert pyver.replace('.', '').isdigit() and len(pyver.split('.')) == 2
     
-    build_idir = abspath(conf_i['build']['idir'])
-    build_idir_parent = ospath.dirname(build_idir)
+    proj_dir = abspath(conf_i['build']['proj_dir'])
+    proj_dir_parent = ospath.dirname(proj_dir)
     #   该值相当于 `_apply_config:vars:srcdir`
-    # # del build_idir
+    # # del proj_dir
     
     # --------------------------------------------------------------------------
     
     # assign conf_i to conf_o
     conf_o['app_name'] = conf_i['app_name']
     conf_o['app_version'] = conf_i['app_version']
-    conf_o['icon'] = abspath(conf_i['icon'] or 'template/python.ico')
+    conf_o['description'] = conf_i['description']
     conf_o['author'] = conf_i['author']
     
-    conf_o['build']['idir'] = abspath(conf_i['build']['idir'])
-    conf_o['build']['odir'] = abspath(conf_i['build']['odir'].format(
+    conf_o['build']['proj_dir'] = abspath(conf_i['build']['proj_dir'])
+    conf_o['build']['dist_dir'] = abspath(conf_i['build']['dist_dir'].format(
         app_name=conf_i['app_name'],
         app_name_lower=conf_i['app_name'].lower().replace(' ', '_'),
         app_version=conf_i['app_version']
     ))
+    conf_o['build']['icon'] = abspath(
+        conf_i['build']['icon'] or 'template/python.ico'
+    )
     conf_o['build']['readme'] = abspath(conf_i['build']['readme'])
     conf_o['build']['module_paths'] = [
-        relpath(p, build_idir_parent) for p in conf_i['build']['module_paths']
+        relpath(p, proj_dir_parent) for p in conf_i['build']['module_paths']
     ]
     conf_o['build']['attachments'] = {
         abspath(k): v for (k, v) in conf_i['build']['attachments'].items()
     }
+    conf_o['build']['enable_console'] = conf_i['build']['enable_console']
     
     conf_o['build']['target'] = conf_i['build']['target']
     conf_o['build']['target']['file'] = relpath(
-        conf_i['build']['target']['file'], build_idir_parent
+        conf_i['build']['target']['file'], proj_dir_parent
     )
     
     conf_o['build']['required'] = conf_i['build']['required']
@@ -121,35 +125,38 @@ def process_pyproject(pyproj_file):
     # run conf_o
     # lk.logp(conf_o)
     _apply_config(conf_o['app_name'], **conf_o['build'], icon=conf_o['icon'])
+    dumps(conf_o, conf_o['build']['dist_dir'] + '/build/manifest.json')
 
 
-def _apply_config(app_name, idir, odir, target, required,
-                  readme='', module_paths=None, attachments=None, **misc):
+def _apply_config(app_name, proj_dir, dist_dir, target, required,
+                  module_paths=None, attachments=None, **misc):
     """
     
     Args:
         app_name (str)
-        idir (str): 文件夹的绝对路径或相对路径. 路径分隔符使用正斜杠
-        odir (str): 'output directory'. 建议填 '../dist/xxx' (xxx 为你要发布的项
-            目的名字, 可以带上版本号. 例如 '../dist/ufs_testcases_0.1.0')
+        proj_dir (str): abspath of dir
+        dist_dir (str): abspath of dir
         target (dict)
-        readme (str): str endswith '.md'. 读我文档, 请确保后缀是 .md (markdown
-            格式的文件), 本程序会把它转换为 html 格式并放在 odir 目录下
         module_paths (list): see `_create_launcher()`
-        attachments (dict): 其他要加入的目录
+        attachments (dict): 附件和资产
         required (dict)
         **misc:
-            icon
+            readme: abspath *.md|*.html|*.pdf|...
+            icon: abspath *.icon
+            enable_console: True|False
     """
     # adjust args
     if module_paths is None: module_paths = []
     if attachments is None: attachments = {}
+    readme = misc.get('readme', '')
     
     # precheck
-    _precheck_args(idir, odir, readme, attachments, required['python_version'])
+    _precheck_args(
+        proj_dir, dist_dir, readme, attachments, required['python_version']
+    )
     
     # if output dirs not exist, create them
-    rootdir, srcdir = odir, f'{odir}/src'
+    rootdir, srcdir = dist_dir, f'{dist_dir}/src'
     #   'root directory', 'source code directory'
     filesniff.force_create_dirpath(srcdir)
     
@@ -157,18 +164,18 @@ def _apply_config(app_name, idir, odir, target, required,
     
     dirs_to_compile = []
     
-    dirs_to_compile.extend(_copy_sources(idir, srcdir))
+    dirs_to_compile.extend(_copy_sources(proj_dir, srcdir))
     
     dirs_to_compile.extend(_copy_assets(attachments, srcdir))
     
     if GlobalConf.create_checkup_tool:
-        os.mkdir(f'{odir}/build')
-        _copy_checkup_tool(f'{odir}/build')
+        _copy_checkup_tool(f'{dist_dir}/build')
     
     _create_launcher(
         app_name, misc.get('icon'), target, rootdir,
         extend_sys_paths=module_paths,
-        enable_venv=required['enable_venv']
+        enable_venv=required['enable_venv'],
+        enable_console=misc.get('enable_console', True),
     )
     
     if readme:
@@ -184,12 +191,11 @@ def _apply_config(app_name, idir, odir, target, required,
         #   you can comment this line to remain .py files for debugging purpose
     
     if required['enable_venv'] and GlobalConf.create_venv_shell:
-        os.mkdir(f'{odir}/venv')
         _copy_venv(required['venv'], f'{rootdir}/venv',
                    required['python_version'],
                    include_tkinter=True)
-        #   include_tkinter: 考虑到我们的 bootloader 有用到 tkinter 的 msgbox, 所
-        #   以这里选用有 tkinter 模块的版本
+        #   include_tkinter: 考虑到我们的 bootloader 有用到 tkinter 的 msgbox,
+        #   所以这里选用有 tkinter 模块的版本
     
     m, n = ospath.split(rootdir)
     lk.logt("[I2501]", f'See distributed project at \n\t"{m}:0" >> {n}')
@@ -197,21 +203,22 @@ def _apply_config(app_name, idir, odir, target, required,
 
 # ------------------------------------------------------------------------------
 
-def _precheck_args(idir, odir, readme, attachments, pyversion):
-    assert ospath.exists(idir)
+def _precheck_args(proj_dir, dist_dir, readme, attachments, pyversion):
+    assert ospath.exists(proj_dir)
     
-    if not odir.isascii():
+    if not dist_dir.isascii():
         lk.loga('警告: 您的输出路径包含有中文字符, 这可能导致软件启动失败! '
-                '(特别在使用虚拟环境的情况下)', odir)
-    if ospath.exists(odir):
-        if os.listdir(odir):
+                '(特别在使用虚拟环境的情况下)', dist_dir)
+    if ospath.exists(dist_dir):
+        if os.listdir(dist_dir):
             if input(
                     '警告: 要打包的目录已存在!\n'
                     '您是否确认清空目标目录以重构: "{}"\n'
-                    '请注意确认删除后内容无法恢复! (y/n): '.format(odir)
+                    '请注意确认删除后内容无法恢复! (y/n): '.format(dist_dir)
             ).lower() == 'y':
-                shutil.rmtree(odir)
-                # os.mkdir(odir)  # we'll make dir laterly in `_apply_config`
+                shutil.rmtree(dist_dir)
+                # # os.mkdir(dist_dir)
+                #   we'll recreate dist_dir laterly in `_apply_config`
             else:
                 raise FileExistsError
     
@@ -219,28 +226,28 @@ def _precheck_args(idir, odir, readme, attachments, pyversion):
     
     assert all(map(ospath.exists, attachments.keys()))
     
-    from .checkup import check_pyversion
+    from .checkup.doctor import check_pyversion
     curr_ver, result = check_pyversion(*map(int, pyversion.split('.')))
     assert result is True, \
         f'prebuild 使用的 Python 版本 ({curr_ver}) ' \
         f'不符合目标编译版本 ({pyversion})!'
 
 
-def _copy_sources(idir, srcdir):
-    """ 将 idir 的内容全部拷贝到 srcdir 下.
+def _copy_sources(proj_dir, srcdir):
+    """ 将 proj_dir 的内容全部拷贝到 srcdir 下.
     
     Args:
-        idir: see `main()`
+        proj_dir: see `main()`
         srcdir: 'source code dir'. 传入 `main:vars:srcdir`
     """
-    yield from _copy_assets({idir: 'assets,compile'}, srcdir)
+    yield from _copy_assets({proj_dir: 'assets,compile'}, srcdir)
 
 
 def _copy_assets(attachments, srcdir):
     """
     Args:
-        attachments (dict): {idir: type, ...}
-            idir (str)
+        attachments (dict): {dir_i: type, ...}
+            dir_i (str)
             type (str): 'assets'|'root_folder'|'root_assets'|'tree_folders'
                 |'compile'|'assets,compile,...' (多个值组合时, 用逗号分隔)
         srcdir
@@ -249,32 +256,30 @@ def _copy_assets(attachments, srcdir):
         dirpath
     """
     
-    def copy_tree_excludes_protected_folders(irootdir, orootdir):
+    def copy_tree_excludes_protected_folders(rootdir_i, rootdir_o):
         invalid_pattern = re.compile(r'/(\.|__?)\w+')
         #   e.g. '/.git', '/__pycache__'
         
-        valid_dirs = [(irootdir, orootdir)]  # [(i, o), ...]
-        # FIXME: 1.4.4 版本的 lk-utils.filesniff.findall_dirs 不完善, 无法完
-        #   全地过滤掉需要被排除的文件, 所以我们自定义一个 invalid_pattern 来
-        #   处理
-        for idir in filesniff.findall_dirs(irootdir):
-            if invalid_pattern.search(idir):
+        valid_dirs = [(rootdir_i, rootdir_o)]  # [(i, o), ...]
+        # FIXME: 1.4.4 版本的 lk-utils.filesniff.findall_dirs 不完善, 无法完全地
+        #   过滤掉需要被排除的文件, 所以我们自定义一个 invalid_pattern 来处理
+        for dir_i in filesniff.findall_dirs(rootdir_i):
+            if invalid_pattern.search(dir_i):
                 continue
-            odir = f'{orootdir}/{idir.replace(irootdir + "/", "", 1)}'
-            lk.logax(idir, odir)
-            valid_dirs.append((idir, odir))
+            dir_o = f'{rootdir_o}/{dir_i.replace(rootdir_i + "/", "", 1)}'
+            valid_dirs.append((dir_i, dir_o))
         lk.reset_count()
         
         for (i, o) in valid_dirs:
             filesniff.force_create_dirpath(o)
             for fp, fn in filesniff.find_files(i, fmt='zip'):
-                ifp, ofp = fp, f'{o}/{fn}'
-                shutil.copyfile(ifp, ofp)
+                fp_i, fp_o = fp, f'{o}/{fn}'
+                shutil.copyfile(fp_i, fp_o)
         del valid_dirs
     
-    for idir, type_ in attachments.items():
-        dirname = ospath.basename(idir)
-        odir = f'{srcdir}/{dirname}'
+    for dir_i, type_ in attachments.items():
+        dirname = ospath.basename(dir_i)
+        dir_o = f'{srcdir}/{dirname}'
         
         ''' pyswitch
         from pyswitch import switch
@@ -290,28 +295,29 @@ def _copy_assets(attachments, srcdir):
         
         if 'assets' in type_:
             if 'compile' in type_:
-                copy_tree_excludes_protected_folders(idir, odir)
+                copy_tree_excludes_protected_folders(dir_i, dir_o)
             else:
-                shutil.copytree(idir, odir)
-        elif 'root_folder' in type_:
-            os.mkdir(odir)
+                shutil.copytree(dir_i, dir_o)
+        elif 'only_folder' in type_:
+            os.mkdir(dir_o)
         elif 'root_assets' in type_:
-            for fp, fn in filesniff.find_files(idir, fmt='zip'):
-                shutil.copyfile(fp, f'{odir}/{fn}')
-            for dp, dn in filesniff.find_dirs(idir, fmt='zip'):
-                os.mkdir(f'{odir}/{dn}')
-        elif 'tree_folders' in type_:
-            for dp, dn in filesniff.findall_dirs(idir, fmt='zip'):
-                os.mkdir(dp.replace(idir, odir, 1))
+            for fp, fn in filesniff.find_files(dir_i, fmt='zip'):
+                shutil.copyfile(fp, f'{dir_o}/{fn}')
+            for dp, dn in filesniff.find_dirs(dir_i, fmt='zip'):
+                os.mkdir(f'{dir_o}/{dn}')
+        elif 'only_folders' in type_:
+            for dp, dn in filesniff.findall_dirs(dir_i, fmt='zip'):
+                os.mkdir(dp.replace(dir_i, dir_o, 1))
         
         if 'compile' in type_:
-            yield odir
+            yield dir_o
 
 
 def _copy_checkup_tool(buildir):
     # buildir: 'build (noun.) directory'
-    shutil.copyfile('checkup.py', f'{buildir}/checkup.py')
-    shutil.copyfile('pretty_print.py', f'{buildir}/pretty_print.py')
+    if not ospath.exists(buildir): os.mkdir(buildir)
+    shutil.copyfile('checkup/doctor.py', f'{buildir}/doctor.py')
+    shutil.copyfile('checkup/pretty_print.py', f'{buildir}/pretty_print.py')
 
 
 def _copy_venv(src_venv_dir, dst_venv_dir, pyversion, include_tkinter=False):
@@ -324,20 +330,20 @@ def _copy_venv(src_venv_dir, dst_venv_dir, pyversion, include_tkinter=False):
             编译器, 以及 src_venv_dir 所用的 Python 版本一致 (修订号可以不一样), 
             否则 _compile_py_files 编译出来的 .pyc 文件无法运行!
         include_tkinter: 默认的 embed python 安装包是不带 tkinter 模块的, 如果需
-            要, 则会使用一个修改过的 embed python 安装包 (修改只涉及复制了 tkinter
-            相关的模块到原生 embed python 安装目录内)
+            要, 则会使用一个修改过的 embed python 安装包 (修改只涉及复制了
+            tkinter 相关的模块到原生 embed python 安装目录内)
     
     Notes:
         1. 本函数使用了 embed_python 独立安装包的内容, 而非简单地把 src_venv_dir
            拷贝到打包目录, 这是因为 src_venv_dir 里面的 Python 是不可独立使用的.
-           也就是说, 在一个没有安装过 Python 的用户电脑上, 调用 src_venv_dir 下的
-           Python 编译器将失败! 所以我们需要的是一个嵌入版的 Python (在 Python 官
-           网下载带有 "embed" 字样的压缩包, 并解压, 我在 pyportable_installer 项目
-           下已经准备了一份)
-        2. 出于性能和成本考虑, 您不必提供有效 src_venv_dir 参数, 即您可以给该参数
-           传入一个空字符串, 这样本函数会仅创建虚拟环境的框架 (dst_venv_dir), 并让
-           '{dst_venv_dir}/site-packages' 留空. 稍后您可以手动地复制, 或剪切所需
-           的依赖到 '{dst_venv_dir}/site-packages'
+           也就是说, 在一个没有安装过 Python 的用户电脑上, 调用 src_venv_dir 下
+           的 Python 编译器将失败! 所以我们需要的是一个嵌入版的 Python (在
+           Python 官网下载带有 "embed" 字样的压缩包, 并解压, 我在
+           pyportable_installer 项目下已经准备了一份)
+        2. 出于性能和成本考虑, 您不必提供有效 src_venv_dir 参数, 即您可以给该参
+           数传入一个空字符串, 这样本函数会仅创建虚拟环境的框架 (dst_venv_dir),
+           并让 '{dst_venv_dir}/site-packages' 留空. 稍后您可以手动地复制, 或剪
+           切所需的依赖到 '{dst_venv_dir}/site-packages'
            
     Results:
         copy source dir to target dir:
@@ -365,7 +371,8 @@ def _copy_venv(src_venv_dir, dst_venv_dir, pyversion, include_tkinter=False):
 
 
 def _create_launcher(app_name, icon, target, rootdir,
-                     extend_sys_paths=None, enable_venv=True):
+                     extend_sys_paths=None, enable_venv=True,
+                     enable_console=True):
     """ 创建启动器.
     
     Args:
@@ -389,18 +396,18 @@ def _create_launcher(app_name, icon, target, rootdir,
         引导装载程序位于: '{rootdir}/src/bootloader.pyc'
         
         1. 二者的体积都非常小
-        2. 启动器本质上是一个带有自定义图标的 bat 脚本. 它指定了 Python 编译器的路
-           径和 bootloader 的路径, 通过调用编译器执行 bootloader.pyc
+        2. 启动器本质上是一个带有自定义图标的 bat 脚本. 它指定了 Python 编译器的
+           路径和 bootloader 的路径, 通过调用编译器执行 bootloader.pyc
         3. bootloader 主要完成了以下两项工作:
             1. 向 sys.path 中添加当前工作目录和自定义的模块目录
-            2. 对主脚本加了一个 try catch 结构, 以便于捕捉主程序报错时的信息, 并以
-               tkinter 弹窗的形式给用户. 这样就摆脱了控制台打印的需要, 使我们的软
-               件表现得更像是一款软件
+            2. 对主脚本加了一个 try catch 结构, 以便于捕捉主程序报错时的信息, 并
+               以 tkinter 弹窗的形式给用户. 这样就摆脱了控制台打印的需要, 使我们
+               的软件表现得更像是一款软件
     
     Notes:
         1. 启动器在调用主脚本 (main:args:main_script) 之前, 会通过 `os.chdir` 切
-           换到主脚本所在的目录, 这样我们项目源代码中的所有相对路径, 相对引用都能正
-           常工作
+           换到主脚本所在的目录, 这样我们项目源代码中的所有相对路径, 相对引用都
+           能正常工作
     
     References:
         template/launch_by_system.bat
@@ -429,8 +436,8 @@ def _create_launcher(app_name, icon, target, rootdir,
         TARGET_KWARGS=str(target['kwargs']),
     )
     dumps(code, f'{rootdir}/src/{bootloader_name}.py')
-    #   这里生成的是 .py 文件. 我们会在 `_apply_config` 的后期做编译工作, 届时会将
-    #   它转换为 .pyc, 并删除 .py 文件. (见 `_apply_config:#:compile '{srcdir}
+    #   这里生成的是 .py 文件. 我们会在 `_apply_config` 的后期做编译工作, 届时会
+    #   将它转换为 .pyc, 并删除 .py 文件. (见 `_apply_config:#:compile '{srcdir}
     #   /bootloader.py'`)
     
     # --------------------------------------------------------------------------
@@ -441,8 +448,8 @@ def _create_launcher(app_name, icon, target, rootdir,
     if enable_venv:  # suggest
         template = loads('template/launch_by_venv.bat')
     else:
-        # 注意: 这个不太安全, 因为我们不能确定用户系统安装默认的 Python 版本是否与
-        # 当前编译的 pyc 版本相同.
+        # 注意: 这个不太安全, 因为我们不能确定用户系统安装默认的 Python 版本是否
+        # 与当前编译的 pyc 版本相同.
         template = loads('template/launch_by_system.bat')
     code = template.format(
         LAUNCHER=f'{bootloader_name}.pyc'
@@ -464,50 +471,52 @@ def _create_launcher(app_name, icon, target, rootdir,
     thread = Thread(
         target=generate_exe,
         args=(bat_file, f'{rootdir}/{launcher_name}.exe', icon,
-              '/x64', 'invisible')
+              '/x64', '' if enable_console else '/invisible')
     )
     thread.start()
     # thread.join()
 
 
-def _create_readme(ifile: str, distdir):
-    ofile = f'{distdir}/{ospath.basename(ifile)}'
-    shutil.copyfile(ifile, ofile)
+def _create_readme(file_i: str, distdir):
+    file_o = f'{distdir}/{ospath.basename(file_i)}'
+    shutil.copyfile(file_i, file_o)
 
 
-def _compile_py_files(idir, recursive=True):
+def _compile_py_files(dir_i, recursive=True):
     """
     References:
         https://blog.csdn.net/weixin_38314865/article/details/90443135
     """
-    compile_dir(idir, maxlevels=None if recursive else 0, quiet=1)
-    #   maxlevels: None 编译全部文件 (包含子文件夹); 0 只编译当前目录下的 py 文件
-    #   quiet: 1 表示只在有错误时向控制台打印信息
+    compile_dir(dir_i, maxlevels=10 if recursive else 0, quiet=1)
+    #   maxlevels: int. 指定遍历的深度, 最小为 0 (0 只编译当前目录下的 py 文件).
+    #       注意该值在 Python 3.8 下默认是 10, 在 Python 3.9 下默认是 None. 为了
+    #       能够在 3.8 及以下正常工作, 所以我用了 10.
+    #   quiet: 1 表示只在有错误时向控制台打印信息.
     
-    for fp, fn in filesniff.findall_files(idir, suffix='.pyc', fmt='zip'):
+    for fp, fn in filesniff.findall_files(dir_i, suffix='.pyc', fmt='zip'):
         #   fp: 'filepath', fn: 'filename', e.g. 'xxx.cpython-38.pyc'
-        ifp = fp
-        ofp = f'{fp}/../../{fn.split(".", 1)[0]}.pyc'
+        fp_i = fp
+        fp_o = f'{fp}/../../{fn.split(".", 1)[0]}.pyc'
         #   1. 第一个 '../' 表示自身所在的目录, 第二个 '../' 指向上一级目录
-        #   2. 不能直接使用 fn, 因为 fn 的后缀是 '.cpython-38.pyc', 会导致 Python
-        #      的导入语法不能正常工作 (提示 "ImportError: No module named
+        #   2. 不能直接使用 fn, 因为 fn 的后缀是 '.cpython-38.pyc', 会导致
+        #      Python 的导入语法不能正常工作 (提示 "ImportError: No module named
         #      'model'"), 为了解决此问题, 将后缀改为 '.pyc' 即可.
-        shutil.move(ifp, ofp)
+        shutil.move(fp_i, fp_o)
 
 
-def _cleanup_py_files(idir, recursive=True):
+def _cleanup_py_files(dir_i, recursive=True):
     if recursive:
         # delete __pycache__ folders (the folders are empty)
         for dp in filesniff.findall_dirs(  # dp: 'dirpath'
-                idir, suffix='__pycache__',
+                dir_i, suffix='__pycache__',
                 exclude_protected_folders=False
         ):
             os.rmdir(dp)
-    
+        
         # and delete .py files
-        for fp in filesniff.findall_files(idir, suffix='.py'):
+        for fp in filesniff.findall_files(dir_i, suffix='.py'):
             os.remove(fp)
     else:
-        os.rmdir(f'{idir}/__pycache__')
-        for fp in filesniff.find_files(idir, suffix='.py'):
+        os.rmdir(f'{dir_i}/__pycache__')
+        for fp in filesniff.find_files(dir_i, suffix='.py'):
             os.remove(fp)